# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kosova Estate",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/faqs": {
            "get": {
                "summary": "Retrieve a list of FAQs",
                "operationId": "2b5fb58f5e0deb541848dd201a46becb",
                "responses": {
                    "200": {
                        "description": "A list of FAQs"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "summary": "Create a new FAQ",
                "operationId": "18c6ac5147188ad557f2c4cd0f21718b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "question": {
                                        "description": "FAQ question",
                                        "type": "string",
                                        "example": "What is the return policy?"
                                    },
                                    "answer": {
                                        "description": "FAQ answer",
                                        "type": "string",
                                        "example": "Our return policy is 30 days."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "FAQ created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/faqs/{id}": {
            "put": {
                "summary": "Update an existing FAQ",
                "operationId": "6fd89df2092fae177a5897f61deecbc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the FAQ to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answer": {
                                        "description": "Updated answer for the FAQ",
                                        "type": "string",
                                        "example": "Our return policy is 30 days."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "FAQ updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/properties/filter": {
            "get": {
                "summary": "Filter properties based on query parameters",
                "operationId": "c48e9866382172e6da9f191aebfd622d",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query for property description or address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "propertyType",
                        "in": "query",
                        "description": "Type of property (sale or rent)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "propertyCategory",
                        "in": "query",
                        "description": "Category of property (residential or business)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered properties list"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    }
}